requirements.txt
Faker~=10.0.0
pytest==6.2.5,<7
python-dotenv~=0.19.2
requests~=2.26.0
selenium~=4.0.0
